Return a context on creation of a machine
so that the caller can monitor the process state and fails if the VM
exits before it should.

`monitor` of the vm process is only implemented in qemu so the context is only
useful in qemu. In the rest of the cases, ctx is the same one passed to
`Create`.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>