Run SSH commands by copying a script
instead of inline. This is to avoid issues with quotes not being escaped
and such. E.g. when trying a command like this:

```
source /etc/os-release; echo $VERSION
```

the `$VERSION` is not escaped and since we wrap the whole command in a

`/bin/sh -c ""`

it gets expanded so VERSION is evaluated before `source` is run.
Instead of fighting to escape quotes for bash/sh from Go, we simply
write the desired command in a script (temp file) and run that.

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>