Add DetachCD to machine interface
although it's only implemented by VBox. That's because we no longer have
access to the global Machine object in tests and test that are using
VBox used to call DetachCD on the machine directly.

Qemu default boot order is "disk first" so we don't need to eject the
cdrom (which fails anyway because there "resource is busy" while booted
with the live cd).

Signed-off-by: Dimitris Karakasilis <dimitris@karakasilis.me>